DT-1759 Threshold is larger than an int

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/indexer/config/IndexBuildProperties.kt
@@ -5,4 +5,4 @@ import org.springframework.boot.context.properties.ConstructorBinding
 
 @ConstructorBinding
 @ConfigurationProperties("index-build")
-data class IndexBuildProperties(val completeThreshold: Int)
+data class IndexBuildProperties(val completeThreshold: Long)
